+++ components/common/cpp/include/ftl/cuda_operators.hpp
@@ -240,7 +240,7 @@ inline __host__ __device__ float2 normalize(float2 v)
 // floor
 inline __host__ __device__ float2 floor(const float2 v)
 {
-    return make_float2(floor(v.x), floor(v.y));
+    return make_float2(floorf(v.x), floorf(v.y));
 }
 
 // reflect
@@ -252,7 +252,7 @@ inline __host__ __device__ float2 reflect(float2 i, float2 n)
 // absolute value
 inline __host__ __device__ float2 fabs(float2 v)
 {
-	return make_float2(fabs(v.x), fabs(v.y));
+	return make_float2(fabsf(v.x), fabsf(v.y));
 }
 
 inline __device__ __host__ int2 sign(float2 f) { 
@@ -423,7 +423,7 @@ inline __host__ __device__ float3 normalize(float3 v)
 // floor
 inline __host__ __device__ float3 floor(const float3 v)
 {
-    return make_float3(floor(v.x), floor(v.y), floor(v.z));
+    return make_float3(floorf(v.x), floorf(v.y), floorf(v.z));
 }
 
 // reflect
@@ -435,7 +435,7 @@ inline __host__ __device__ float3 reflect(float3 i, float3 n)
 // absolute value
 inline __host__ __device__ float3 fabs(float3 v)
 {
-	return make_float3(fabs(v.x), fabs(v.y), fabs(v.z));
+	return make_float3(fabsf(v.x), fabsf(v.y), fabsf(v.z));
 }
 
 inline __device__ __host__ int3 sign(float3 f) { 
@@ -567,13 +567,13 @@ inline __host__ __device__ float4 normalize(float4 v)
 // floor
 inline __host__ __device__ float4 floor(const float4 v)
 {
-    return make_float4(floor(v.x), floor(v.y), floor(v.z), floor(v.w));
+    return make_float4(floorf(v.x), floorf(v.y), floorf(v.z), floorf(v.w));
 }
 
 // absolute value
 inline __host__ __device__ float4 fabs(float4 v)
 {
-	return make_float4(fabs(v.x), fabs(v.y), fabs(v.z), fabs(v.w));
+	return make_float4(fabsf(v.x), fabsf(v.y), fabsf(v.z), fabsf(v.w));
 }
 
 // int3 functions
